remove decoration phase from batch
