+++ flink-libraries/flink-table/src/test/scala/org/apache/flink/table/expressions/utils/ExpressionTestBase.scala
@@ -24,7 +24,7 @@ import java.util.concurrent.Future
 import org.apache.calcite.rex.RexNode
 import org.apache.calcite.sql.`type`.SqlTypeName._
 import org.apache.calcite.sql2rel.RelDecorrelator
-import org.apache.calcite.tools.{Programs, RelBuilder, RuleSet}
+import org.apache.calcite.tools.{Programs, RelBuilder}
 import org.apache.flink.api.common.TaskInfo
 import org.apache.flink.api.common.accumulators.Accumulator
 import org.apache.flink.api.common.functions._
@@ -66,10 +66,10 @@ abstract class ExpressionTestBase {
     context._2.getTypeFactory)
   private val optProgram = Programs.ofRules(FlinkRuleSets.DATASET_OPT_RULES)
 
-  private def hepPlanner(ruleSet: RuleSet) = {
+  private def hepPlanner = {
     val builder = new HepProgramBuilder
     builder.addMatchOrder(HepMatchOrder.BOTTOM_UP)
-    val it = ruleSet.iterator()
+    val it = FlinkRuleSets.DATASET_NORM_RULES.iterator()
     while (it.hasNext) {
       builder.addRuleInstance(it.next())
     }
@@ -185,7 +185,7 @@ abstract class ExpressionTestBase {
 
     // normalize
     val normalizedPlan = if (FlinkRuleSets.DATASET_NORM_RULES.iterator().hasNext) {
-      val planner = hepPlanner(FlinkRuleSets.DATASET_NORM_RULES)
+      val planner = hepPlanner
       planner.setRoot(decorPlan)
       planner.findBestExp
     } else {
@@ -194,16 +194,7 @@ abstract class ExpressionTestBase {
 
     // create DataSetCalc
     val flinkOutputProps = converted.getTraitSet.replace(DataSetConvention.INSTANCE).simplify()
-    val optimizedPlan = optProgram.run(context._2.getPlanner, normalizedPlan, flinkOutputProps)
-
-    // decorate
-    val dataSetCalc = if (FlinkRuleSets.DATASET_DECO_RULES.iterator().hasNext) {
-      val planner = hepPlanner(FlinkRuleSets.DATASET_DECO_RULES)
-      planner.setRoot(optimizedPlan)
-      planner.findBestExp
-    } else {
-      optimizedPlan
-    }
+    val dataSetCalc = optProgram.run(context._2.getPlanner, normalizedPlan, flinkOutputProps)
 
     // extract RexNode
     val calcProgram = dataSetCalc
